Class Lookups to determine import class target from item meta.
Class lookups will be performed when importing items depending
on items metadata and the registered class lookups.

When a lookup matches, the item will be imported in the target
class. Otherwise, it goes in the class selected by the user.

By default, the verbatim taoQtiItem extension does not register
any ClassLookups and behaves as it behaved before.